fix(#1490): point LLMTR at the moved typed-object path
#1445 moved the shared utils to src/shared; #1266 merged after it still
imported @/lib/utils/typed-object. Each was green alone, so main went red on
the merge: six error TS plus an unresolved-import failure in knip.
